Add service layer for Sight
