+++ JavaServer/src/main/java/com/project/piss/SightController.java
@@ -2,6 +2,8 @@ package com.project.piss;
 
 
 import com.project.piss.models.Sight;
+import com.project.piss.services.SightService;
+import com.project.piss.services.SightServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -28,23 +30,21 @@ public class SightController {
 
 
     @Autowired
-    private SightRepository repository;
+    private SightServiceImpl sightService;
 
 
-    //implement exception
     @GetMapping("/v1/sights/{id}/retrieve")
     public Sight findAll(@PathVariable("id") Long id) {
-        return repository.findById(id)
-                .orElseThrow(RuntimeException::new);
+        return sightService.findById(id);
     }
+
 
     //add optional arguments
     @GetMapping("/v1/sights/retrieve")
     public List<Sight> findAll() {
-        List<Sight> sights = repository.findAll();
-        sights.forEach(sight -> sight.setPicture_path(URL + sight.getPicture_path()));
-        return sights;
+        return sightService.findAll();
     }
+
 
     //remove all arguments and add class on their place
     @PostMapping(value = "/v1/sights/create", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -59,67 +59,20 @@ public class SightController {
                         @RequestParam("latitude") double latitude,
                         @RequestParam("category") String category
     ) throws IOException {
-        String newImageName = generateNewImageName(picture);
-        saveUploadedFile(picture, newImageName);
-        final double INITIAL_RATING = 0;
-        final boolean IS_PENDING_DEFAULT_STATUS = true;
-        Sight sight = new Sight(name, description, INITIAL_RATING, newImageName, workingTimeFrom, workingTimeTo, price, address,
-                longitude, latitude, category, IS_PENDING_DEFAULT_STATUS);
-        return repository.save(sight);
+        return sightService.save(name, description, picture, workingTimeFrom, workingTimeTo, price, address,
+                longitude, latitude, category);
     }
 
-    private String generateNewImageName(MultipartFile file) {
-
-        String generatedName = generateName();
-        String fileExtension = getFileExtension(file);
-        String newImageName = generatedName + "." + fileExtension;
-        return newImageName;
-    }
-
-    private String generateName() {
-        return UUID.randomUUID().toString();
-    }
-
-
-    //not handled exception
-    private String getFileExtension(MultipartFile file) {
-        String fileName = file.getOriginalFilename();
-        int INDEX_OF_FILE_EXTENSION = 1;
-        String REGEX_FOR_MATCH_FILE_EXTENSION = "\\.(?=[^\\.]+$)";
-        return fileName.split(REGEX_FOR_MATCH_FILE_EXTENSION)[INDEX_OF_FILE_EXTENSION];
-    }
-
-
-    private void saveUploadedFile(MultipartFile file, String pathOfImage) throws IOException {
-        if (!file.isEmpty()) {
-            byte[] bytes = file.getBytes();
-            String UPLOADED_FOLDER = "src/main/resources/static/images/";
-            Path path = Paths.get(UPLOADED_FOLDER + pathOfImage);
-            Files.write(path, bytes);
-        }
-    }
 
     @DeleteMapping("/v1/sights/{id}/delete")
-    public ResponseEntity<?> deleteSight(@PathVariable("id") long id) {
-
-        return repository.findById(id)
-                .map(sight -> {
-                    delete(sight.getPicture_path(), id);
-                    return ResponseEntity.ok().build();
-                }).orElse(ResponseEntity.notFound().build());
-    }
-
-    private void delete(String imageName, long id) {
-
-        File file = new File(URL + imageName);
-        file.delete();
-        repository.deleteById(id);
+    public ResponseEntity<?> deleteSight(@PathVariable("id") Long id) {
+        return sightService.deleteSight(id);
     }
 
 
 //    @PatchMapping("/v1/sights/{id}/manage")
 //    public ResponseEntity<Sight> updateSight(@PathVariable("id") long id, @RequestParam("title") String title, @RequestParam("file") MultipartFile file
 //    ) {
-//        return cardServcie.update(id, title, file);
+//        return sightService.update(id, title, file);
 //    }
 }
