+++ JavaServer/src/main/java/com/project/piss/SightRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface SightRepository
+public interface SightRepository
         extends JpaRepository<Sight, Long> {
 }
